refactor(configurator): split preset page modules
