+++ configurator/src/app/pages/presets.rs
@@ -9,6 +9,9 @@
 //! `enabled` flag, its expansion mirrors `preset_collapsed`, and every row
 //! inside sends the same indexed [`Message`] the Iced slot view sent.
 
+mod color;
+mod rows;
+
 use relm4::prelude::*;
 use relm4::{adw, gtk};
 
@@ -26,6 +29,8 @@ use crate::models::{
 use super::super::search::{AppSearchSummary, SearchArea};
 use super::super::state::ConfiguratorApp;
 use super::{BuiltPage, PageBuilder, set_selected_blocked, set_text_blocked, validate_u32_range};
+use color::build_color_rows;
+use rows::{SlotBuilder, labels_of, slot_button};
 
 pub(super) fn build(sender: &ComponentSender<ConfiguratorApp>) -> BuiltPage {
     let mut page = PageBuilder::new(sender, TabId::Presets);
@@ -288,301 +293,4 @@ fn build_slot(page: &mut PageBuilder, slot: usize) {
                 .map(|draft| draft.show_status_bar)
         },
     );
-}
-
-/// The slot's color block: a mode chooser with a live preview, then the
-/// rows the chosen mode edits.
-fn build_color_rows(rows: &mut SlotBuilder<'_>, slot: usize) {
-    let mode_row = rows.combo_row(
-        "Color mode",
-        vec![ColorMode::Named, ColorMode::Rgb],
-        vec!["Named".to_string(), "RGB".to_string()],
-        move |app| app.draft.presets.slot(slot).map(|draft| draft.color.mode),
-        Message::PresetColorModeChanged,
-    );
-    let swatch = color_swatch();
-    mode_row.add_suffix(&swatch);
-    // The color the swatch is currently drawing, owned by the binding that
-    // draws it: a new draw function is what redraws the widget, so it is only
-    // installed when the color actually moved.
-    let mut shown: Option<[f64; 3]> = None;
-    rows.page.bind(move |app, _search| {
-        let color = app
-            .draft
-            .presets
-            .slot(slot)
-            .and_then(|draft| draft.color.preview_color())
-            .map(|color| [color.r, color.g, color.b]);
-        if shown == color {
-            return;
-        }
-        shown = color;
-        match color {
-            Some(rgb) => swatch.set_draw_func(move |_, context, width, height| {
-                context.set_source_rgb(rgb[0], rgb[1], rgb[2]);
-                context.rectangle(0.0, 0.0, f64::from(width), f64::from(height));
-                // A failed fill only leaves the swatch blank.
-                let _ = context.fill();
-            }),
-            // A slot whose color resolves to nothing draws nothing.
-            None => swatch.set_draw_func(|_, _, _, _| {}),
-        }
-    });
-
-    let named = NamedColorOption::list();
-    let named_labels = labels_of(&named, NamedColorOption::label);
-    let named_row = rows.combo_row(
-        "Named color",
-        named,
-        named_labels,
-        move |app| {
-            app.draft
-                .presets
-                .slot(slot)
-                .map(|draft| draft.color.selected_named)
-        },
-        Message::PresetNamedColorSelected,
-    );
-    rows.visible_when(&named_row, move |app| {
-        app.draft
-            .presets
-            .slot(slot)
-            .is_some_and(|draft| draft.color.mode == ColorMode::Named)
-    });
-
-    let custom_name = rows.entry_row_validated(
-        "Custom color name",
-        move |app| {
-            app.draft
-                .presets
-                .slot(slot)
-                .map(|draft| draft.color.name.clone())
-        },
-        move |slot, value| Message::PresetTextChanged(slot, PresetTextField::ColorName, value),
-        move |app| {
-            let color = &app.draft.presets.slot(slot)?.color;
-            if color.mode != ColorMode::Named {
-                return None;
-            }
-            (color.preview_color().is_none() && !color.name.trim().is_empty())
-                .then(|| "Unknown color name.".to_string())
-        },
-    );
-    rows.visible_when(&custom_name, move |app| {
-        app.draft.presets.slot(slot).is_some_and(|draft| {
-            draft.color.mode == ColorMode::Named && draft.color.selected_named_is_custom()
-        })
-    });
-
-    for (component, title) in [
-        (0usize, "Red (0-255)"),
-        (1, "Green (0-255)"),
-        (2, "Blue (0-255)"),
-    ] {
-        let row = rows.entry_row_validated(
-            title,
-            move |app| {
-                app.draft
-                    .presets
-                    .slot(slot)
-                    .and_then(|draft| draft.color.rgb.get(component).cloned())
-            },
-            move |slot, value| Message::PresetColorComponentChanged(slot, component, value),
-            move |app| {
-                let color = &app.draft.presets.slot(slot)?.color;
-                if color.mode != ColorMode::Rgb {
-                    return None;
-                }
-                validate_u32_range(color.rgb.get(component)?, 0, 255)
-            },
-        );
-        rows.visible_when(&row, move |app| {
-            app.draft
-                .presets
-                .slot(slot)
-                .is_some_and(|draft| draft.color.mode == ColorMode::Rgb)
-        });
-    }
-}
-
-/// Row builder for one slot.
-///
-/// Mirrors the [`PageBuilder`] row helpers, but adds rows to the slot's
-/// expander and reads the slot out of the model, so a slot the draft does
-/// not hold simply leaves its rows alone.
-struct SlotBuilder<'a> {
-    page: &'a mut PageBuilder,
-    expander: adw::ExpanderRow,
-    slot: usize,
-}
-
-impl SlotBuilder<'_> {
-    /// A free-text row sending `to_message(slot, text)` on change.
-    fn entry_row(
-        &mut self,
-        title: &str,
-        get: impl Fn(&ConfiguratorApp) -> Option<String> + 'static,
-        to_message: impl Fn(usize, String) -> Message + 'static,
-    ) -> adw::EntryRow {
-        self.entry_row_validated(title, get, to_message, |_app| None)
-    }
-
-    /// A free-text row with live validation: a non-`None` result marks the
-    /// row `.error` and shows the text as its tooltip.
-    fn entry_row_validated(
-        &mut self,
-        title: &str,
-        get: impl Fn(&ConfiguratorApp) -> Option<String> + 'static,
-        to_message: impl Fn(usize, String) -> Message + 'static,
-        validate: impl Fn(&ConfiguratorApp) -> Option<String> + 'static,
-    ) -> adw::EntryRow {
-        let row = adw::EntryRow::builder().title(title).build();
-        let slot = self.slot;
-        let handler = {
-            let sender = self.page.sender();
-            row.connect_changed(move |row| {
-                sender.input(to_message(slot, row.text().to_string()));
-            })
-        };
-        self.expander.add_row(&row);
-        {
-            let row = row.clone();
-            self.page.bind(move |app, _search| {
-                if let Some(value) = get(app) {
-                    // Blocked: the draft owns this text, and a load reporting
-                    // its own value back as a user edit clears that load's
-                    // diagnostics from the status line.
-                    set_text_blocked(&row, &handler, &value);
-                }
-                set_row_error(&row, validate(app));
-            });
-        }
-        row
-    }
-
-    /// A single-choice row sending `to_message(slot, value)` on selection.
-    fn combo_row<O>(
-        &mut self,
-        title: &str,
-        values: Vec<O>,
-        labels: Vec<String>,
-        get: impl Fn(&ConfiguratorApp) -> Option<O> + 'static,
-        to_message: impl Fn(usize, O) -> Message + 'static,
-    ) -> adw::ComboRow
-    where
-        O: Copy + PartialEq + 'static,
-    {
-        let label_refs: Vec<&str> = labels.iter().map(String::as_str).collect();
-        let row = adw::ComboRow::builder()
-            .title(title)
-            .model(&gtk::StringList::new(&label_refs))
-            .build();
-        let slot = self.slot;
-        let handler: SignalHandlerId = {
-            let sender = self.page.sender();
-            let values = values.clone();
-            row.connect_selected_notify(move |row| {
-                if let Some(value) = values.get(row.selected() as usize) {
-                    sender.input(to_message(slot, *value));
-                }
-            })
-        };
-        self.expander.add_row(&row);
-        {
-            let row = row.clone();
-            self.page.bind(move |app, _search| {
-                let Some(current) = get(app) else {
-                    return;
-                };
-                if let Some(index) = values.iter().position(|value| *value == current) {
-                    // Blocked: the draft chose this, and reporting it back as
-                    // a user pick clears the status line a load just wrote.
-                    set_selected_blocked(&row, &handler, index as u32);
-                }
-            });
-        }
-        row
-    }
-
-    /// A Default/On/Off row for one of the slot's override fields.
-    fn override_row(
-        &mut self,
-        title: &str,
-        field: PresetToggleField,
-        get: impl Fn(&ConfiguratorApp) -> Option<OverrideOption> + 'static,
-    ) -> adw::ComboRow {
-        let values = OverrideOption::list();
-        let labels = labels_of(&values, OverrideOption::label);
-        self.combo_row(title, values, labels, get, move |slot, value| {
-            Message::PresetToggleOptionChanged(slot, field, value)
-        })
-    }
-
-    /// Binds a row's visibility to a model condition.
-    fn visible_when(
-        &mut self,
-        row: &impl IsA<gtk::Widget>,
-        visible: impl Fn(&ConfiguratorApp) -> bool + 'static,
-    ) {
-        let row = row.clone().upcast::<gtk::Widget>();
-        self.page.bind(move |app, _search| {
-            let value = visible(app);
-            if row.is_visible() != value {
-                row.set_visible(value);
-            }
-        });
-    }
-}
-
-/// A flat header button for one of the slot's actions.
-fn slot_button(page: &PageBuilder, label: &str, tooltip: &str, message: Message) -> gtk::Button {
-    let button = gtk::Button::builder()
-        .label(label)
-        .tooltip_text(tooltip)
-        .valign(gtk::Align::Center)
-        .css_classes(["flat"])
-        .build();
-    let sender = page.sender();
-    button.connect_clicked(move |_| sender.input(message.clone()));
-    button
-}
-
-/// A read-only preview of the slot's resolved color.
-///
-/// Blank until its binding installs a draw function carrying the color, so
-/// the widget holds no state the binding has to encode and decode.
-fn color_swatch() -> gtk::DrawingArea {
-    gtk::DrawingArea::builder()
-        .content_width(24)
-        .content_height(24)
-        .valign(gtk::Align::Center)
-        .css_classes(["card"])
-        .build()
-}
-
-fn set_row_error(row: &adw::EntryRow, error: Option<String>) {
-    let has_error_class = row.has_css_class("error");
-    match error {
-        Some(message) => {
-            if !has_error_class {
-                row.add_css_class("error");
-            }
-            if row.tooltip_text().as_deref() != Some(message.as_str()) {
-                row.set_tooltip_text(Some(&message));
-            }
-        }
-        None => {
-            if has_error_class {
-                row.remove_css_class("error");
-                row.set_tooltip_text(None);
-            }
-        }
-    }
-}
-
-fn labels_of<O>(values: &[O], label: impl Fn(&O) -> &'static str) -> Vec<String> {
-    values
-        .iter()
-        .map(|value| label(value).to_string())
-        .collect()
 }
